+++ packages/react-food-ya/src/context/themeContext.jsx
@@ -1,0 +1,17 @@
+/* eslint-disable react/prop-types */
+/* eslint-disable react/destructuring-assignment */
+/* eslint-disable react/jsx-no-constructed-context-values */
+import React, { createContext, useState } from 'react';
+
+const ThemeContext = createContext(undefined);
+
+function ThemeProvider(props) {
+  const [theme, setTheme] = useState('light');
+
+  return (
+    <ThemeContext.Provider value={{ theme, setTheme }}>
+      {props.children}
+    </ThemeContext.Provider>
+  );
+}
+export { ThemeContext, ThemeProvider };
